requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langchain-postgres==0.0.1
langchain-text-splitters==0.0.1
langchain-core==0.1.20
openai==1.12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pypdf==4.0.1
